+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Iterator.sum`
 - `Iterator.position`
 - `Iterator.fold`
+- `convert.ToString`
 
 ### Changed